+++ packages/graphql/src/schema/question.ts
@@ -250,6 +250,27 @@ export const Element = ElementRef.implement({
   }),
 })
 
+export interface IQuestionOrElementInstance {
+  questionInstance?: IQuestionInstance | null
+  elementInstance?: DB.ElementInstance | null
+}
+export const QuestionOrElementInstanceRef =
+  builder.objectRef<IQuestionOrElementInstance>('QuestionOrElementInstance')
+export const QuestionOrElementInstance = QuestionOrElementInstanceRef.implement(
+  {
+    fields: (t) => ({
+      questionInstance: t.expose('questionInstance', {
+        type: QuestionInstanceRef,
+        nullable: true,
+      }),
+      elementInstance: t.expose('elementInstance', {
+        type: ElementInstanceRef,
+        nullable: true,
+      }),
+    }),
+  }
+)
+
 export interface IQuestionInstance extends DB.QuestionInstance {
   evaluation?: IInstanceEvaluationOLD
 }