+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [3.1.0-rc.16](https://github.com/uzh-bf/klicker-uzh/compare/v3.1.0-rc.15...v3.1.0-rc.16) (2024-03-06)
+
+
+### Bug Fixes
+
+* **packages/graphql:** ensure builds work for new updateQuestionInstances ([eff3ace](https://github.com/uzh-bf/klicker-uzh/commit/eff3ace669c7bb3915b4e98e70eff8aeccadb71b))
+
+## [3.1.0-rc.15](https://github.com/uzh-bf/klicker-uzh/compare/v3.1.0-rc.14...v3.1.0-rc.15) (2024-03-06)
+
+
+### Bug Fixes
+
+* ensure updateQuestionInstances works on old session and new practice quiz and microlearning, remove MicroSession ([#4059](https://github.com/uzh-bf/klicker-uzh/issues/4059)) ([913051e](https://github.com/uzh-bf/klicker-uzh/commit/913051eb490864c78c9af63bb76e9f214fa43b0e))
+
 ## [3.1.0-rc.14](https://github.com/uzh-bf/klicker-uzh/compare/v3.1.0-rc.13...v3.1.0-rc.14) (2024-03-05)